reorganizacion en paquetes
