--- src/main/java/edu/fiuba/algo3/model/Score.java
@@ -1,9 +1,0 @@
-package edu.fiuba.algo3.model;
-
-//TODO: implementar la logica de Score
-public class Score {
-    // cuenta los enemigos matados
-    // cuenta los creditos del jugador
-
-
-}
