+++ src/main/java/edu/fiuba/algo3/model/excepciones/CreditosInsuficientesError.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.model.exceptions;
+package edu.fiuba.algo3.model.excepciones;
 
 public class CreditosInsuficientesError extends RuntimeException{
 }
