+++ src/main/java/edu/fiuba/algo3/model/mapa/Camino.java
@@ -1,7 +1,7 @@
-package edu.fiuba.algo3.model.map;
+package edu.fiuba.algo3.model.mapa;
 
-import edu.fiuba.algo3.model.Jugador;
-import edu.fiuba.algo3.model.enemy.Enemigo;
+import edu.fiuba.algo3.model.jugador.Jugador;
+import edu.fiuba.algo3.model.enemigos.Enemigo;
 
 import java.util.ArrayList;
