+++ src/main/java/edu/fiuba/algo3/model/mapa/Pasarela.java
@@ -1,8 +1,8 @@
-package edu.fiuba.algo3.model.map;
+package edu.fiuba.algo3.model.mapa;
 
-import edu.fiuba.algo3.model.defense.Defensa;
-import edu.fiuba.algo3.model.Jugador;
-import edu.fiuba.algo3.model.enemy.Enemigo;
+import edu.fiuba.algo3.model.defensa.Defensa;
+import edu.fiuba.algo3.model.jugador.Jugador;
+import edu.fiuba.algo3.model.enemigos.Enemigo;
 
 import java.util.ArrayList;
